backend/carbon-footprint/requirements.txt
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
pillow==10.4.0
google-generativeai==0.8.3
requests==2.32.3
pytesseract==0.3.10